Move skel into the core module of tklife.
Refactor tests to better reflect the changes made to the package's
structure.

Add exception messages to SkeletonMixin widget creation to help
developers locate errors faster.

Update api documentation to reflect changes in package structure.

Update to issue #8.
