+++ docs/source/api.rst
@@ -1,12 +1,35 @@
 API Reference
 =============
 
-Skel
-----
+.. automodule:: tklife
+    :special-members: __version__
 
-.. automodule:: tklife.skel
+tklife.core
+-----------
+
+.. automodule:: tklife.core
     :members:
     :show-inheritance:
     :inherited-members:
     :special-members: __before_init__, __after_init__, __after_widgets__, __getitem__, __setitem__, __getattr__, __setattr__
+    :member-order: bysource
 
+tklife.menu
+-----------
+
+.. automodule:: tklife.menu
+    :members:
+    :show-inheritance:
+    :inherited-members:
+    :special-members: __before_init__, __after_init__, __after_widgets__, __getitem__, __setitem__, __getattr__, __setattr__
+    :member-order: bysource
+
+tklife.event
+------------
+
+.. automodule:: tklife.event
+    :members:
+    :show-inheritance:
+    :inherited-members:
+    :special-members: __add__
+    :member-order: bysource
