+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "tklife"
-version = "2.3.0-dev4"
+version = "2.4.0-dev0"
 description = "Make Tk life easier"
 authors = ["Cory Laughlin <corylcomposinger@gmail.com>"]
 license = "MIT"
